Move ZoomSlider into separate file
