+++ ui/v2.5/src/components/List/ListViewOptions.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef, useState } from "react";
 import Mousetrap from "mousetrap";
-import { Button, Dropdown, Form, Overlay, Popover } from "react-bootstrap";
+import { Button, Dropdown, Overlay, Popover } from "react-bootstrap";
 import { DisplayMode } from "src/models/list-filter/types";
 import { useIntl } from "react-intl";
 import { Icon } from "../Shared/Icon";
@@ -11,53 +11,7 @@ import {
   faTags,
   faThLarge,
 } from "@fortawesome/free-solid-svg-icons";
-
-interface IZoomSelectProps {
-  minZoom: number;
-  maxZoom: number;
-  zoomIndex: number;
-  onChangeZoom: (v: number) => void;
-}
-
-export const ZoomSelect: React.FC<IZoomSelectProps> = ({
-  minZoom,
-  maxZoom,
-  zoomIndex,
-  onChangeZoom,
-}) => {
-  useEffect(() => {
-    Mousetrap.bind("+", () => {
-      if (zoomIndex !== undefined && zoomIndex < maxZoom) {
-        onChangeZoom(zoomIndex + 1);
-      }
-    });
-    Mousetrap.bind("-", () => {
-      if (zoomIndex !== undefined && zoomIndex > minZoom) {
-        onChangeZoom(zoomIndex - 1);
-      }
-    });
-
-    return () => {
-      Mousetrap.unbind("+");
-      Mousetrap.unbind("-");
-    };
-  });
-
-  return (
-    <Form.Control
-      className="zoom-slider"
-      type="range"
-      min={minZoom}
-      max={maxZoom}
-      value={zoomIndex}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-        onChangeZoom(Number.parseInt(e.currentTarget.value, 10));
-        e.preventDefault();
-        e.stopPropagation();
-      }}
-    />
-  );
-};
+import { ZoomSelect } from "./ZoomSlider";
 
 interface IListViewOptionsProps {
   zoomIndex?: number;
